Made it work.
Fixed it (again), after doing some breaking structural changes (prior to the
initial commit.)

Moved the module-global Node class to internally within the HuffmanTree class.

Split decode() into decode() and decode_char(), but realised that decode_char
must be able to access a variable number of encoded text chars to decode a
single char of decoded text, so now it's passed the index, and returns the
count that decode should shift.

Instead of using a dictionary for char -> encoded_string, the tree has a
leaves structure with char -> leaf_node; this way the same tree structure is
used for encoding and decoding text. It may be good to investigate some kind
of bi-directional dictionary for size purposes.